HAI-3611 Modularize E2E test helpers and reduce ALLU logins
Split monolithic e2e/_setup.ts (634 lines) into focused helper modules
under e2e/helpers/: test-data, auth, hanke, johtoselvitys, kaivuilmoitus,
allu. _setup.ts is now a barrel re-export for backwards compatibility.

Extract repeated inline ALLU login+search blocks from 5 spec files into
alluLogin/alluSearchApplication helpers. Make alluLogin idempotent by
checking the ALLU home page for an active session before re-authenticating,
reducing redundant logins within a single test.

Fix flaky tests: increase toPass timeout in kaivuilmoitus_toiminnallinen_kunto
from 120s to 240s, and press Escape after the last combobox selection in
createAndFillHankeForm to prevent the dropdown from blocking the next click.
