+++ .env.e2e.example
@@ -9,10 +9,8 @@ TA_HAITATON_DEV=<fill in>
 TA_HAITATON_STAGING=<fill in>
 TA_HAITATON_PROD="#" # Production environment not used in E2E tests
 
-TA_ALLU_DEV="#"
-TA_ALLU_TESTING=<fill in>
-TA_ALLU_STAGING="#"
-TA_ALLU_PROD="#"
+TA_ALLU_LOGIN=<fill in>
+TA_ALLU_HOME=<fill in>
 
 TA_ASIAKAS_USERNAME="Philip Auvinen"
 TA_ASIAKAS_EMAIL="haitaton.e2e.philip.auvinen@test.invalid"
